Moved slim select setup to separated file
